requirements.txt
aniso8601==8.0.0
apispec==5.1.0
astroid==2.4.2
autopep8==1.5.5
bidict==0.21.3
certifi==2020.6.20
chardet==3.0.4
charset-normalizer==2.0.4
click==7.1.1
Flask==1.1.2
flask-apispec==0.11.0
Flask-Cors==3.0.10
Flask-JWT-Extended==3.24.1
Flask-RESTful==0.3.8
idna==2.10
isort==5.7.0
itsdangerous==1.1.0
Jinja2==2.11.2
json-excel-converter==1.3.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
marshmallow==3.13.0
mccabe==0.6.1
numpy==1.21.2
pika==1.2.0
Pillow==8.3.2
pycodestyle==2.6.0
pycryptodome==3.9.8
PyJWT==1.7.1
pylint==2.6.0
pymongo==3.12.0
PyMySQL==1.0.2
python-dotenv==0.15.0
python-engineio==4.2.1
python-socketio==5.4.0
pytz==2019.3
pyzmq==22.2.1
redis==3.5.3
requests==2.23.0
rq==1.7.0
six==1.14.0
toml==0.10.2
urllib3==1.25.11
webargs==8.0.0
websocket-client==1.2.1
Werkzeug==1.0.1
wrapt==1.12.1